RF-18469 Abstract away the Validator
